Refactor project structure, add  ESLint configurations, and reorganize app files. Added CI-CD pipeline
