+++ .github/templates/pull-request-template.md
@@ -1,0 +1,61 @@
+## Full Stack Pull Request
+
+### 🔄 Scope of Changes
+- [ ] Client only
+- [ ] Server only
+- [ ] Both client and server
+
+### 📱 Client Changes (if applicable)
+#### Type of Change
+- [ ] Bug fix
+- [ ] New feature
+- [ ] UI/UX improvements
+- [ ] Performance improvements
+- [ ] Other: _____
+
+#### Client Details
+<!-- Describe client-side changes here -->
+
+#### Client Checklist
+- [ ] Tested on iOS
+- [ ] Tested on Android
+- [ ] Added/Updated client tests
+- [ ] Checked responsive layouts
+- [ ] Updated client documentation
+
+### ⚙️ Server Changes (if applicable)
+#### Type of Change
+- [ ] Bug fix
+- [ ] New feature
+- [ ] API modification
+- [ ] Database changes
+- [ ] Other: _____
+
+#### Server Details
+<!-- Describe server-side changes here -->
+
+#### Server Checklist
+- [ ] Added/Updated server tests
+- [ ] Updated API documentation
+- [ ] Added database migrations (if needed)
+- [ ] Checked security implications
+- [ ] Verified API performance
+
+### 🔗 Related Issues
+<!-- Link related issues using #issue-number -->
+
+### 📸 Screenshots
+<!-- Add screenshots of UI changes if applicable -->
+
+### 🧪 Test Coverage
+<!-- Describe how you tested your changes -->
+
+### 📝 Additional Notes
+<!-- Any additional information reviewers should know -->
+
+### ✅ Final Checklist
+- [ ] I have performed a self-review
+- [ ] Code follows project style guidelines
+- [ ] All tests pass
+- [ ] Documentation has been updated
+- [ ] Changes have been tested in a dev environment
