+++ .github/workflows/server-ci-cd.yaml
@@ -1,0 +1,70 @@
+name: Server CI/CD
+on:
+  push:
+    branches: ['**']
+  pull_request:
+    branches:
+      - main
+      - dev
+
+jobs:
+  build-and-test:
+    runs-on: ubuntu-latest
+    defaults:
+      run:
+        working-directory: ./server
+
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v4
+        with:
+          fetch-depth: 0
+
+      - name: Setup .NET
+        uses: actions/setup-dotnet@v3
+        with:
+          dotnet-version: '8.0.x'
+
+      - name: Restore dependencies
+        run: dotnet restore
+
+      - name: Build
+        run: dotnet build --configuration Release --no-restore
+
+      - name: Run tests
+        run: dotnet test --no-build --verbosity normal --configuration Release
+
+      # Optional: Code coverage analysis
+      - name: Generate code coverage report
+        run: dotnet test --no-build /p:CollectCoverage=true /p:CoverletOutputFormat=opencover
+
+  deploy:
+    needs: build-and-test
+    runs-on: ubuntu-latest
+    if: github.ref == 'refs/heads/main' || github.ref == 'refs/heads/dev'
+
+    defaults:
+      run:
+        working-directory: ./server
+
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v4
+
+      - name: Setup .NET
+        uses: actions/setup-dotnet@v3
+        with:
+          dotnet-version: '8.0.x'
+
+      - name: Publish
+        run: |
+          dotnet restore
+          dotnet publish -c Release -o ./publish
+
+    #   - name: Deploy to Azure
+    #     uses: azure/webapps-deploy@v2
+    #     if: github.ref == 'refs/heads/main' # Only deploy from main branch
+    #     with:
+    #       app-name: your-azure-app-name # Replace with your Azure app name
+    #       package: ./server/publish
+    #       publish-profile: ${{ secrets.AZURE_WEBAPP_PUBLISH_PROFILE }}
