+++ client/.prettierrc
@@ -1,0 +1,8 @@
+{
+  "semi": false,
+  "tabWidth": 2,
+  "printWidth": 100,
+  "singleQuote": false,
+  "trailingComma": "none",
+  "jsxBracketSameLine": false
+}
