+++ client/App.tsx
@@ -1,0 +1,20 @@
+import { StatusBar } from "expo-status-bar"
+import { StyleSheet, Text, View } from "react-native"
+
+export default function App() {
+  return (
+    <View style={styles.container}>
+      <Text>Hello olivia</Text>
+      <StatusBar style="auto" />
+    </View>
+  )
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: "#fff",
+    alignItems: "center",
+    justifyContent: "center"
+  }
+})
