+++ client/package.json
@@ -1,0 +1,48 @@
+{
+  "name": "h-com",
+  "version": "1.0.0",
+  "main": "expo-router/entry",
+  "scripts": {
+    "start": "expo start",
+    "android": "expo start --android",
+    "ios": "expo start --ios",
+    "web": "expo start --web",
+    "type-check": "tsc --noEmit",
+    "lint": "eslint .",
+    "lint:fix": "eslint --fix . --ext .ts,.tsx --fix",
+    "format": "prettier --write './**/*.{js,jsx,ts,tsx,css,md,json}' --config ./.prettierrc"
+  },
+  "dependencies": {
+    "@hookform/resolvers": "^3.9.1",
+    "@react-native-async-storage/async-storage": "^2.1.0",
+    "@react-native/eslint-config": "^0.77.0",
+    "@rneui/themed": "^4.0.0-rc.8",
+    "@supabase/supabase-js": "^2.47.10",
+    "aes-js": "^3.1.2",
+    "dotenv": "^16.4.7",
+    "eslint-plugin-react-native": "^5.0.0",
+    "expo": "~52.0.18",
+    "expo-constants": "~17.0.3",
+    "expo-image-picker": "~16.0.4",
+    "expo-linking": "~7.0.3",
+    "expo-router": "4.0.11",
+    "expo-secure-store": "~14.0.0",
+    "expo-status-bar": "~2.0.0",
+    "prettier": "^3.4.2",
+    "react": "18.3.1",
+    "react-hook-form": "^7.54.1",
+    "react-native": "0.76.5",
+    "react-native-get-random-values": "^1.11.0",
+    "react-native-image-picker": "^7.2.3",
+    "react-native-safe-area-context": "4.12.0",
+    "react-native-screens": "~4.1.0",
+    "zod": "^3.24.1"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.25.2",
+    "@types/aes-js": "^3.1.4",
+    "@types/react": "~18.3.12",
+    "typescript": "^5.3.3"
+  },
+  "private": true
+}
