+++ client/src/app/(community)/map/_layout.tsx
@@ -1,0 +1,9 @@
+import { Stack } from "expo-router"
+
+export default function MapLayout() {
+  return (
+    <Stack>
+      <Stack.Screen name="index" options={{ headerShown: false }} />
+    </Stack>
+  )
+}
