+++ client/src/app/(community)/profile/_layout.tsx
@@ -1,0 +1,9 @@
+import { Stack } from "expo-router"
+
+export default function ProfileLayout() {
+  return (
+    <Stack>
+      <Stack.Screen name="index" options={{ headerShown: false }} />
+    </Stack>
+  )
+}
