+++ client/src/app/assets/adaptive-icon.png
