+++ client/src/app/assets/icon.png
