+++ client/src/app/assets/splash-icon.png
